# specs/001-backend-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Beautique Store Backend API
  description: |
    RESTful API for Beautique Store e-commerce platform.

    ## Authentication
    - Public endpoints: No authentication required
    - Admin endpoints (/api/admin/*): JWT Bearer token required

    ## Payment Workflow
    Phase 1 uses manual payment verification via WhatsApp.
    No payment gateway integration.
  version: 1.0.0
  contact:
    name: Beautique Store
    email: admin@beautique.com
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://beautique-api.hf.space
    description: Production server (Hugging Face Spaces)
...
paths:
...
  /api/products:
    get:
      tags: [Products]
      summary: List all products
      description: Returns paginated list of active products with optional filters
      operationId: listProducts
      parameters:
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/ProductCategory'
          description: Filter by category
        - name: search
          in: query
          schema:
            type: string
          description: Search in name and description
        - name: minPrice
          in: query
          schema:
            type: number
            minimum: 0
          description: Minimum price filter
        - name: maxPrice
          in: query
          schema:
            type: number
            maximum: 1000000
          description: Maximum price filter
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 50
          description: Items per page
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
  /api/products/{id}:
    get:
      tags: [Products]
      summary: Get product details
      description: Returns complete product information by ID
      operationId: getProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/products/bestsellers:
    get:
      tags: [Products]
      summary: Get bestseller products
      description: Returns products marked as bestsellers
      operationId: getBestsellers
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 20
      responses:
        '200':
          description: List of bestseller products
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
  /api/orders:
    post:
      tags: [Orders]
      summary: Create new order
      description: |
        Creates a new order with customer details and cart items.
        Returns generated Order ID (format: BQ-YYYYMMDD-XXX).
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/orders/track:
    get:
      tags: [Orders]
      summary: Track order
      description: |
        Returns order details when Order ID and phone number match.
        Returns "Order not found" if either doesn't match (security).
      operationId: trackOrder
      parameters:
        - name: order_id
          in: query
          required: true
          schema:
            type: string
          description: Order ID (e.g., BQ-20260118-001)
        - name: phone
          in: query
          required: true
          schema:
            type: string
          description: Phone number used when ordering
      responses:
        '200':
          description: Order tracking information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderTrackingResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================
  # ADMIN ENDPOINTS (JWT REQUIRED)
  # ============================================
...
  /api/admin/products:
    post:
      tags: [Admin Products]
      summary: Create product
      description: Creates a new product
      operationId: createProduct
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/admin/products/{id}:
    put:
      tags: [Admin Products]
      summary: Update product
      description: Updates an existing product
      operationId: updateProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Admin Products]
      summary: Delete product
      description: Soft-deletes a product (sets is_active=false)
      operationId: deleteProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Product deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/admin/orders:
    get:
      tags: [Admin Orders]
      summary: List all orders
      description: Returns paginated list of orders with filters
      operationId: listOrders
      security:
        - BearerAuth: []
      parameters:
        - name: payment_status
          in: query
          schema:
            $ref: '#/components/schemas/PaymentStatus'
        - name: order_status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/orders/{id}:
    get:
      tags: [Admin Orders]
      summary: Get order details
      description: Returns complete order information
      operationId: getOrder
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProductCategory:
      type: string
      enum:
        - Maxi
        - Lehanga Choli
        - Long Shirt
        - Shalwar Kameez
        - Gharara
    PaymentMethod:
      type: string
      enum:
        - Easypaisa
        - Meezan Bank
        - International Bank
    PaymentStatus:
      type: string
      enum:
        - Pending
        - Paid
        - Verified
    OrderStatus:
      type: string
      enum:
        - Received
        - Processing
        - Ready
        - Delivered
        - Cancelled
    DeliveryStatus:
      type: string
      enum:
        - Not Started
        - In Progress
        - Out for Delivery
        - Delivered

    # ============================================
    # PRODUCT SCHEMAS
    # ============================================
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 3
          maxLength: 255
        category:
          $ref: '#/components/schemas/ProductCategory'
        price:
          type: number
          minimum: 0
          maximum: 1000000
        description:
          type: string
          minLength: 10
          maxLength: 2000
        images:
          type: array
          items:
            type: string
            format: uri
          minItems: 1
          maxItems: 10
        sizes:
          type: array
          items:
            type: string
            enum: [S, M, L, XL, XXL]
          minItems: 1
        colors:
          type: array
          items:
            type: string
          minItems: 1
        is_bestseller:
          type: boolean
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductCreate:
      type: object
      required:
        - name
        - category
        - price
        - description
        - images
        - sizes
        - colors
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 255
        category:
          $ref: '#/components/schemas/ProductCategory'
        price:
          type: number
          minimum: 0
          maximum: 1000000
        description:
          type: string
          minLength: 10
          maxLength: 2000
        images:
          type: array
          items:
            type: string
            format: uri
          minItems: 1
          maxItems: 10
        sizes:
          type: array
          items:
            type: string
            enum: [S, M, L, XL, XXL]
          minItems: 1
        colors:
          type: array
          items:
            type: string
          minItems: 1
        is_bestseller:
          type: boolean
          default: false
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 255
        category:
          $ref: '#/components/schemas/ProductCategory'
        price:
          type: number
          minimum: 0
          maximum: 1000000
        description:
          type: string
          minLength: 10
          maxLength: 2000
        images:
          type: array
          items:
            type: string
            format: uri
          minItems: 1
          maxItems: 10
        sizes:
          type: array
          items:
            type: string
            enum: [S, M, L, XL, XXL]
          minItems: 1
        colors:
          type: array
          items:
            type: string
          minItems: 1
        is_bestseller:
          type: boolean
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer

    # ============================================
    # ORDER SCHEMAS
    # ============================================
    OrderItem:
      type: object
      required:
        - product_id
        - name
        - size
        - color
        - quantity
        - price
      properties:
        product_id:
          type: string
          format: uuid
        name:
          type: string
        size:
          type: string
        color:
          type: string
        quantity:
          type: integer
          minimum: 1
        price:
          type: number
          minimum: 0
    OrderCreate:
      type: object
      required:
        - customer_name
        - phone
        - whatsapp
        - address
        - city
        - items
        - payment_method
      properties:
        customer_name:
          type: string
          minLength: 3
        phone:
          type: string
          pattern: '^[\+]?[(]?[0-9]{1,4}[)]?[-\s\.]?[(]?[0-9]{1,4}[)]?[-\s\.]?[0-9]{1,9}$'
        whatsapp:
          type: string
          pattern: '^[\+]?[(]?[0-9]{1,4}[)]?[-\s\.]?[(]?[0-9]{1,4}[)]?[-\s\.]?[0-9]{1,9}$'
        email:
          type: string
          format: email
        address:
          type: string
          minLength: 10
        city:
          type: string
          minLength: 2
        country:
          type: string
        notes:
          type: string
          maxLength: 500
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          minItems: 1
        payment_method:
          $ref: '#/components/schemas/PaymentMethod'
    OrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          example: BQ-20260118-001
        customer_name:
          type: string
        total_amount:
          type: number
        payment_status:
          $ref: '#/components/schemas/PaymentStatus'
        order_status:
          $ref: '#/components/schemas/OrderStatus'
        created_at:
          type: string
          format: date-time
    OrderTrackingResponse:
      type: object
      properties:
        order_id:
          type: string
        customer_name:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        total_amount:
          type: number
        payment_method:
          $ref: '#/components/schemas/PaymentMethod'
        payment_status:
          $ref: '#/components/schemas/PaymentStatus'
        order_status:
          $ref: '#/components/schemas/OrderStatus'
        delivery_status:
          $ref: '#/components/schemas/DeliveryStatus'
        estimated_delivery:
          type: string
          format: date
        tracking_notes:
          type: string
        order_date:
          type: string
          format: date
    OrderDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
        customer_name:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        total_amount:
          type: number
        payment_method:
          $ref: '#/components/schemas/PaymentMethod'
        payment_status:
          $ref: '#/components/schemas/PaymentStatus'
        order_status:
          $ref: '#/components/schemas/OrderStatus'
        delivery_status:
          $ref: '#/components/schemas/DeliveryStatus'
        estimated_delivery:
          type: string
          format: date
        tracking_notes:
          type: string
        order_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              order_id:
                type: string
              customer_name:
                type: string
              total_amount:
                type: number
              payment_status:
                $ref: '#/components/schemas/PaymentStatus'
              order_status:
                $ref: '#/components/schemas/OrderStatus'
              created_at:
                type: string
                format: date-time
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer

    # ============================================
    # ADMIN SCHEMAS
    # ============================================
...
    Error:
      type: object
      properties:
        detail:
          type: string
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string
  responses:
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Invalid or expired token
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Resource not found
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'